clang-tidy and clang-format

+++ src/testing.h
@@ -1,12 +1,8 @@
-#include <array>
 #include <cstdint>
-#include <string>
-#include <utility>
-#include <vector>
 
 #include "bitboard.h"
 
 void run_perft_test(int max_draft);
 template <side_t side>
 auto perft_search(BitBoard &board, int iter) -> uint64_t;
-void test_puzzles(int count);
+void test_puzzles(size_t count);
